Move controlledPromise to core-utils to deduplicate
