+++ src/components/file-browser/useFileDownloadCommand.test.ts
@@ -1,7 +1,7 @@
+import { controlledPromise } from '@terra-ui-packages/core-utils';
 import { act, renderHook } from '@testing-library/react';
 import FileBrowserProvider, { FileBrowserFile } from 'src/libs/ajax/file-browser-providers/FileBrowserProvider';
 import { reportError } from 'src/libs/error';
-import { controlledPromise } from 'src/testing/test-utils';
 
 import { useFileDownloadCommand } from './useFileDownloadCommand';
