+++ src/components/file-browser/useFileDownloadUrl.test.ts
@@ -1,6 +1,6 @@
+import { controlledPromise } from '@terra-ui-packages/core-utils';
 import { act, renderHook } from '@testing-library/react';
 import FileBrowserProvider, { FileBrowserFile } from 'src/libs/ajax/file-browser-providers/FileBrowserProvider';
-import { controlledPromise } from 'src/testing/test-utils';
 
 import { reportError } from '../../libs/error';
 import { useFileDownloadUrl } from './useFileDownloadUrl';
